Moved the ploting functions which was more complex than expected
